+++ src/novedades/index.ts
@@ -1,0 +1,6 @@
+// src/novedades/index.ts
+export * from './types'
+export * from './novedades.repository'
+export * from './novedad.detector'
+export * from './novedades.service'
+export * from './admin.handler'